Add configuration tests and update CI pipelines
