+++ .github/workflows/ci.yml
@@ -18,12 +18,20 @@ jobs:
       - name: Run pre-commit
         uses: pre-commit/action@v3.0.1
 
-  smoke:
-    name: Smoke
+  tests:
+    name: Tests
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      - name: Run smoke checks
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+      - name: Install dependencies
         run: |
-          bash tests/e2e/test_smoke.sh
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+          pip install -r requirements-dev.txt
+      - name: Run tests
+        run: pytest
