+++ actions/__init__.py
@@ -1,2 +1,1 @@
 """Custom DataHub extensions."""
-
