+++ actions/base64_action/__init__.py
@@ -1,2 +1,1 @@
 """Base64 tokenization action package."""
-
