+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+pythonpath = .
+markers =
+    smoke: Quick smoke-level tests for the base64 action configuration helpers.
